// src/app/domain/models/base-tutor.model.ts
import { Review } from './review.model';
import { Rating } from './rating.model';

export interface BaseTutor {
  id: number;
  name: string;
  imageUrl: string;
  headline: string;
  price?: number;
  likes: number;
  rating: Rating;
  reviews: Review[];
  location: {
    country: string;
    city: string;
  };
    
}